% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{adjust}
\alias{adjust}
\title{Capture adjustment covariates}
\usage{
adjust(spec, ...)
}
\arguments{
\item{spec}{A \code{simtab_spec}.}

\item{...}{Tidyselect expressions for adjustment covariates.}
}
\value{
A modified \code{simtab_spec}.
}
\description{
Capture adjustment covariates
}
\examples{
adjust(simtab(epitabl), age, sex)
}
