% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{as.data.frame.simtab_rbind_tb}
\alias{as.data.frame.simtab_rbind_tb}
\title{Convert rbind_tb to Data Frame}
\usage{
\method{as.data.frame}{simtab_rbind_tb}(x, row.names = NULL, optional = FALSE, tidy = FALSE, ...)
}
\arguments{
\item{x}{An \code{rbind_tb} object.}

\item{row.names}{NULL or a character vector giving the row names for the data frame.}

\item{optional}{Logical. If TRUE, setting row names and converting column names is optional.}

\item{tidy}{Logical. If \code{TRUE}, returns a long-format tidy data frame.}

\item{...}{Additional arguments.}
}
\value{
A data.frame.
}
\description{
Convert rbind_tb to Data Frame
}
\examples{
t1 <- tb(epitabl, sex, diabetes)
t2 <- tb(epitabl, hypertension, diabetes)
as.data.frame(rbind(t1, t2))
}
