% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{describe}
\alias{describe}
\title{Capture variables to describe}
\usage{
describe(spec, ...)
}
\arguments{
\item{spec}{A \code{simtab_spec}.}

\item{...}{Tidyselect expressions for variables to describe.}
}
\value{
A modified \code{simtab_spec}.
}
\description{
Capture variables to describe
}
\examples{
describe(simtab(epitabl), age, sex)
}
