% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_engines}
\alias{list_engines}
\title{List registered SimtablR engines}
\usage{
list_engines()
}
\value{
A character vector of registered engine names.
}
\description{
List registered SimtablR engines
}
\examples{
list_engines()
}
\seealso{
\code{\link[=register_engine]{register_engine()}}
}
