% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{missingness}
\alias{missingness}
\title{Set missing-data policy}
\usage{
missingness(spec, display = NULL, denominator = NULL, model = NULL)
}
\arguments{
\item{spec}{A \code{simtab_spec}.}

\item{display}{Whether to display missingness rows.}

\item{denominator}{Missing-data denominator policy, \code{"available"} or
\code{"complete"}.}

\item{model}{Missing-data model policy, \code{"drop"}, \code{"fail"}, or
\code{"explicit"}.}
}
\value{
A modified \code{simtab_spec}.
}
\description{
Set missing-data policy
}
\examples{
missingness(simtab(epitabl), display = TRUE, denominator = "available")
}
