% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{overall}
\alias{overall}
\title{Toggle the overall column}
\usage{
overall(spec, value = TRUE)
}
\arguments{
\item{spec}{A \code{simtab_spec}.}

\item{value}{\code{TRUE} or \code{FALSE}.}
}
\value{
A modified \code{simtab_spec}.
}
\description{
Toggle the overall column
}
\examples{
overall(simtab(epitabl), FALSE)
}
