% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.R
\name{print.regtab}
\alias{print.regtab}
\title{Print Method for regtab Results}
\usage{
\method{print}{regtab}(x, ...)
}
\arguments{
\item{x}{A data.frame returned by \code{regtab()}.}

\item{...}{Additional arguments passed to \code{print()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for regtab Results
}
