% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{print.simtab_rbind_tb}
\alias{print.simtab_rbind_tb}
\title{Print Method for simtab_rbind_tb Objects}
\usage{
\method{print}{simtab_rbind_tb}(x, digits = NULL, ...)
}
\arguments{
\item{x}{A \code{simtab_rbind_tb} object.}

\item{digits}{Minimum number of significant digits to be printed.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for simtab_rbind_tb Objects
}
\examples{
t1 <- tb(epitabl, sex, diabetes)
t2 <- tb(epitabl, hypertension, diabetes)
print(rbind(t1, t2))
}
