% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{print.tb}
\alias{print.tb}
\title{Print Method for tb Objects}
\usage{
\method{print}{tb}(x, digits = NULL, ...)
}
\arguments{
\item{x}{A \code{tb} object.}

\item{digits}{Number of decimal places to display.}

\item{...}{Additional arguments.}
}
\value{
No return value, called for side effects
}
\description{
Prints a formatted table with flags and attributes.
}
