% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{SimtablR-package}
\alias{SimtablR}
\alias{SimtablR-package}
\title{SimtablR: Easy Publication-Ready Tables and Regression Analysis}
\description{
Streamlines the creation of descriptive frequency tables ('Table 1'), diagnostic test accuracy evaluations (sensitivity, specificity, predictive values), and multi-outcome regression summaries. Features a grammar for publication-ready epidemiological tables built on design-aware effect measures, prevalence and odds ratio calculations, and seamless integration with 'flextable' for exporting results to 'Microsoft Word' and 'PowerPoint'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MatheusTG-14/SimtablR}
  \item \url{https://MatheusTG-14.github.io/SimtablR/}
  \item Report bugs at \url{https://github.com/MatheusTG-14/SimtablR/issues}
}

}
\author{
\strong{Maintainer}: Matheus Trabuco Gonzalez \email{matheustrabucogonzalez@gmail.com}

Authors:
\itemize{
  \item Matheus Trabuco Gonzalez \email{matheustrabucogonzalez@gmail.com}
}

}
\keyword{internal}
