% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklists.R
\name{stard}
\alias{stard}
\title{Build a STARD reporting checklist}
\usage{
stard(x, ...)
}
\arguments{
\item{x}{A SimtablR diagnostic or ROC result, report, or spec.}

\item{...}{Ignored.}
}
\value{
A \code{simtab_checklist} object.
}
\description{
Build a STARD reporting checklist
}
\examples{
d <- diag_test(epitabl, poc_hstn_positive, adjudicated_acs,
               positive = "Yes", test_positive = "Positive")
stard(d)
}
