% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{stratify}
\alias{stratify}
\title{Capture a stratifying variable}
\usage{
stratify(spec, by)
}
\arguments{
\item{spec}{A \code{simtab_spec}.}

\item{by}{A single data-masked variable.}
}
\value{
A modified \code{simtab_spec}.
}
\description{
Capture a stratifying variable
}
\examples{
stratify(describe(simtab(epitabl), age, sex), adjudicated_acs)
}
