% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklists.R
\name{strobe}
\alias{strobe}
\alias{as_flextable.simtab_checklist}
\title{Build a STROBE reporting checklist}
\usage{
strobe(x, ...)

\method{as_flextable}{simtab_checklist}(x, ...)
}
\arguments{
\item{x}{A \code{simtab_checklist} object.}

\item{...}{Ignored.}
}
\value{
A \code{simtab_checklist} object.
}
\description{
Build a STROBE reporting checklist
}
\examples{
res <- tb(epitabl, sex, diabetes)
strobe(res)
}
