% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{style}
\alias{style}
\title{Set or update a SimtablR style}
\usage{
style(x, journal)
}
\arguments{
\item{x}{A SimtablR object.}

\item{journal}{A single style or journal preset name.}
}
\value{
A modified SimtablR object.
}
\description{
Set or update a SimtablR style
}
\examples{
style(simtab(epitabl), "default")
}
