% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{tidy.simtab_result}
\alias{tidy.simtab_result}
\alias{tidy.simtab_spec}
\title{Tidy a SimtablR result}
\usage{
\method{tidy}{simtab_result}(x, ...)

\method{tidy}{simtab_spec}(x, ...)
}
\arguments{
\item{x}{A \code{simtab_result} or \code{simtab_spec}.}

\item{...}{Ignored.}
}
\value{
A long data.frame. For \code{table1}, this matches
\code{as.data.frame(x, tidy = TRUE)}.
}
\description{
Tidy a SimtablR result
}
\examples{
res <- tb(epitabl, sex, diabetes)
generics::tidy(res)
}
