% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.R
\name{as.data.frame.diag_test}
\alias{as.data.frame.diag_test}
\title{Convert diag_test Object to Data Frame}
\usage{
\method{as.data.frame}{diag_test}(x, ...)
}
\arguments{
\item{x}{A diag_test object}

\item{...}{Additional arguments (unused)}
}
\value{
A data.frame containing the performance metrics
}
\description{
Convert diag_test Object to Data Frame
}
