% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{as.data.frame.tb}
\alias{as.data.frame.tb}
\title{Convert tb to Data Frame}
\usage{
\method{as.data.frame}{tb}(x, ...)
}
\arguments{
\item{x}{A \code{tb} object.}

\item{...}{Additional arguments.}
}
\value{
A data.frame with the formatted table.
}
\description{
Extracts the data from a tb object as a standard data.frame.
}
