% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tb.R
\name{as_flextable.tb}
\alias{as_flextable.tb}
\title{Convert tb Object to Flextable}
\usage{
\method{as_flextable}{tb}(x, ...)
}
\arguments{
\item{x}{A \code{tb} object.}

\item{...}{Additional arguments passed to \code{flextable::flextable()}.}
}
\value{
A \code{flextable} object.
}
\description{
Convert tb Object to Flextable
}
