% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.R
\name{export_regtab_csv}
\alias{export_regtab_csv}
\title{Export regtab Results to CSV}
\usage{
export_regtab_csv(x, file, ...)
}
\arguments{
\item{x}{A data.frame from \code{regtab()}.}

\item{file}{File path.}

\item{...}{Additional arguments passed to \code{write.csv()}.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Export regtab Results to CSV
}
