% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.R
\name{export_regtab_csv}
\alias{export_regtab_csv}
\title{Export regtab Results to CSV}
\usage{
export_regtab_csv(x, file, ...)
}
\arguments{
\item{x}{A data.frame from regtab}

\item{file}{File path for export}

\item{...}{Additional arguments passed to write.csv}
}
\description{
Export regtab Results to CSV
}
