% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.R
\name{export_regtab_xlsx}
\alias{export_regtab_xlsx}
\title{Export regtab Results to Excel}
\usage{
export_regtab_xlsx(x, file, ...)
}
\arguments{
\item{x}{A data.frame from regtab}

\item{file}{File path for export (.xlsx)}

\item{...}{Additional arguments passed to openxlsx::write.xlsx}
}
\value{
No return value, called for side effects
}
\description{
Requires the 'openxlsx' package
}
