% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regtab.R
\name{print.regtab}
\alias{print.regtab}
\title{Print Method for regtab Results}
\usage{
\method{print}{regtab}(x, ...)
}
\arguments{
\item{x}{A data.frame returned by regtab}

\item{...}{Additional arguments passed to print}
}
\value{
No return value, called for side effects
}
\description{
Print Method for regtab Results
}
